fix: Replaced var type from int to size_t

+++ 9-binary_tree_height.c
@@ -9,8 +9,8 @@
 
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int left_height;
-	int right_height;
+	size_t left_height;
+	size_t right_height;
 
 	if (!tree)
 		return (-1);
